# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v1/articles:
    get:
      tags:
        - Articles
      summary: List Articles
      description: Get a paginated list of Articles with optional filtering and sorting
      operationId: listArticles
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArticleList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Articles
      summary: Create Article
      description: Create a new Article
      operationId: createArticle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateArticleRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Article'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/article_links:
    get:
      tags:
        - ArticleLinks
      summary: List ArticleLinks
      description: Get a paginated list of ArticleLinks with optional filtering and sorting
      operationId: listArticleLinks
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArticleLinkList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - ArticleLinks
      summary: Create ArticleLink
      description: Create a new ArticleLink
      operationId: createArticleLink
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateArticleLinkRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArticleLink'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    Article:
      type: object
      required:
        - id
        - title
        - body
        - status
        - revision
        - metadata
      properties:
        id:
          type: string
          format: uuid
        category_id:
          nullable: true
          type: string
          format: uuid
        title:
          type: string
          maxLength: 200
        body:
          type: string
          maxLength: 20000
        status:
          type: string
        revision:
          type: integer
          format: int32
        published_at:
          nullable: true
          type: string
          format: date-time
        metadata:
          type: object
...
    CreateArticleRequest:
      type: object
      required:
        - title
        - body
        - status
        - revision
        - metadata
      properties:
        category_id:
          nullable: true
          type: string
          format: uuid
        title:
          type: string
          maxLength: 200
        body:
          type: string
          maxLength: 20000
        status:
          type: string
        revision:
          type: integer
          format: int32
        published_at:
          nullable: true
          type: string
          format: date-time
        metadata:
          type: object
...
    ArticleList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Article'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    ArticleLink:
      type: object
      required:
        - id
        - article_id
        - target_module
        - target_type
        - target_id
        - metadata
      properties:
        id:
          type: string
          format: uuid
        article_id:
          type: string
          format: uuid
        target_module:
          type: string
          maxLength: 40
        target_type:
          type: string
          maxLength: 40
        target_id:
          type: string
          format: uuid
        category_key:
          nullable: true
          type: string
          maxLength: 40
        metadata:
          type: object
...
    CreateArticleLinkRequest:
      type: object
      required:
        - article_id
        - target_module
        - target_type
        - target_id
        - metadata
      properties:
        article_id:
          type: string
          format: uuid
        target_module:
          type: string
          maxLength: 40
        target_type:
          type: string
          maxLength: 40
        target_id:
          type: string
          format: uuid
        category_key:
          nullable: true
          type: string
          maxLength: 40
        metadata:
          type: object
...
    ArticleLinkList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ArticleLink'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []